+++ frontend/src/server/fetchMetricsPDF.ts
@@ -1,7 +1,8 @@
 import { handleAppError } from 'app/global-error'
-import { VIEWS_URL } from 'utils/credentials'
+import { API_URL } from 'utils/credentials'
 export const fetchMetricsPDF = async (path: string, fileName: string): Promise<void> => {
-  const response = await fetch(`${VIEWS_URL}owasp/project-health-metrics/${path}`, {
+  const baseUrl = API_URL.split('/api')[0]
+  const response = await fetch(`${baseUrl}/owasp/project-health-metrics/${path}`, {
     method: 'GET',
     headers: {
       accept: 'application/pdf',
@@ -20,7 +21,7 @@ export const fetchMetricsPDF = async (path: string, fileName: string): Promise<v
     const url = window.URL.createObjectURL(blob)
     const link = document.createElement('a')
     link.href = url
-    link.setAttribute('download', fileName)
+    link.setAttribute('download', `${fileName}_${new Date().toISOString().split('T')[0]}.pdf`)
     document.body.appendChild(link)
     link.click()
     document.body.removeChild(link)
